+++ test/src/main/java/org/vertexium/test/GraphTestBase.java
@@ -6912,6 +6912,21 @@ public abstract class GraphTestBase {
         assertEquals(2, vertexPropertyCountByValue.get("Sam").size());
         assertEquals(1L, (long) vertexPropertyCountByValue.get("Sam").get("male"));
         assertEquals(2L, (long) vertexPropertyCountByValue.get("Sam").get("female"));
+
+        q = v1.getExtendedData().query(AUTHORIZATIONS_A_AND_B).limit(0);
+        agg = new TermsAggregation("terms-count", "name");
+        agg.addNestedAggregation(new TermsAggregation("nested", "gender"));
+        assumeTrue("terms aggregation not supported", q.isAggregationSupported(agg));
+        q.addAggregation(agg);
+        aggregationResult = q.extendedDataRows().getAggregationResult("terms-count", TermsResult.class);
+        vertexPropertyCountByValue = nestedTermsBucketToMap(aggregationResult.getBuckets(), "nested");
+
+        assertEquals(2, vertexPropertyCountByValue.size());
+        assertEquals(1, vertexPropertyCountByValue.get("Joe").size());
+        assertEquals(1L, (long) vertexPropertyCountByValue.get("Joe").get("male"));
+        assertEquals(2, vertexPropertyCountByValue.get("Sam").size());
+        assertEquals(1L, (long) vertexPropertyCountByValue.get("Sam").get("male"));
+        assertEquals(2L, (long) vertexPropertyCountByValue.get("Sam").get("female"));
     }
 
     @Test
@@ -8151,6 +8166,7 @@ public abstract class GraphTestBase {
         assertEquals(5, count(graph.getExtendedData(ElementType.VERTEX, "v1", null, AUTHORIZATIONS_A)));
         assertEquals(5, count(graph.getExtendedData(ElementType.VERTEX, null, null, AUTHORIZATIONS_A)));
         assertEquals(5, count(graph.getExtendedData(null, null, null, AUTHORIZATIONS_A)));
+        assertEquals(5, count(v1.getExtendedData()));
         try {
             count(graph.getExtendedData(null, null, "table1", AUTHORIZATIONS_A));
             fail("nulls to the left of a value is not allowed");
@@ -10128,7 +10144,7 @@ public abstract class GraphTestBase {
         getGraph().prepareVertex("v6", VISIBILITY_A)
             .addPropertyValue("k1", "prop1", "aaaaaaaaaaaaaaaaaaaaaa", VISIBILITY_B)
             .addPropertyValue("k2", "prop1", "a", VISIBILITY_B)
-            .addPropertyValue("k3","prop1", "ddddd", VISIBILITY_A)
+            .addPropertyValue("k3", "prop1", "ddddd", VISIBILITY_A)
             .save(AUTHORIZATIONS_A_AND_B);
         getGraph().flush();